requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
pandas>=2.2.0
numpy>=1.26.0
websocket-client>=1.6.4
colorama>=0.4.6
tabulate>=0.9.0
krakenex>=2.2.2
Flask>=3.0.0
